DAGGERLITE-32: Provide an API for full validation with RT.
For more info on API see the docs.

Accompanying changes:
- Log the component from builder creation time, as it was overlooked before.
- Provide unified logging procedure.
- Extract common validation plugin class into a `[:spi-impl]`.
- Fix the kotlin class comparison issue.
- Partially include `api-dynamic-optimized` into testing to check
 if proguard breaks anything.

Use the resulting API for validation in [:testing] instead of
 ad-hoc validation implementation.
The new implementation is based on bootstrapping code generation.
For this to work, an isolated classloader must be used for runtime tests;
 this might slow tests down a bit, as classes need to be loaded every time.
The only newly imposed limitation is that error messages can't be
 grouped across different roots, as RT can't validate multiple roots
 at once.