Api: Rework API for `Yatagan` loaders and reflection setup.
`Yatagan` EP is now defined a single time and sits in `api:public`,
along with the implementation for loading compiled graphs.
It can delegate to reflection backend (via `ImplementationLoader` interface)
 if it is present in the runtime classpath (`api:dynamic`).

Reflection-specific API is now a separate thing, `YataganReflection`,
 that sits in `:api:dynamic`.

`api:compiled` is now an empty artifact left for some degree of compatibility.

With these changes it's now possible to access the Yatagan EP without
 choosing the specific backend, i.e. loader always tries to load
 compiled implementations optionally falling back to reflection,
 if it's present in the runtime classpath.

As this design was implemented with library development in mind, it's
 now very important to keep `Yatagan` loader being able to load
 implementations generated by earlier versions of Yatagan's codegen.
 For that purpose a special `LoaderCompatibilityTest` is introduced.

Yatagan 1.2.0 already broke `create()` loader compatibility, so
 this CL fixes that with a workaround.