Api: Move `Provides.conditionals` API.
Instead of
`@Provides(Conditional(..), Conditional(..))`
it's now required to write `@Provides @Conditional(..) @Conditional(..)`.

That is done to fix IDE's issues with `OptIn`s,
where it could require the user to opt-in to
conditions usage, where none were used.

This also improves consistency with other parts
of the API where `Conditional` is a top-level annotation.

Edits in tests are necessary to separate the api
 classpath, as this CL introduces a breaking change.