Api: Support dagger-compat mode.
This newly introduced mode can now be
enabled with `yatagan.experimental.enableDaggerCompatibility` option.

This mode is designed for yatagan to be a drop-in replacement of
dagger, with *minimal* migration work. And all the necessary
migration actions will keep the code compatible with the
vanilla dagger.

If active, all dagger annotations and types that
have supported functionality in yatagan are recognized, while
yatagan annotations and types can be used interchangeably.

Components can be created with `Dagger<class-name>` generated facades;
`Yatagan.*` entry-points are also available, but only
for classes, that are annotated with `com.yandex.yatagan.Component` annotations.

One can keep both dagger and yatagan annotations on declarations;
the yatagan ones have priority over dagger ones.

RT support: possible, but TBD.